fix: use Double instead of Integer

+++ app/src/main/java/com/example/calcit/commands/DivisionCommand.java
@@ -1,35 +1,35 @@
 package com.example.calcit.commands;
 
 public class DivisionCommand implements OperationCommand {
-    private Integer operand1, operand2,result;
-    public DivisionCommand(Integer operand1, Integer operand2) {
+    private Double operand1, operand2,result;
+    public DivisionCommand(Double operand1, Double operand2) {
         this.operand1 = operand1;
         this.operand2 = operand2;
     }
 
-    public Integer getOperand1() {
+    public Double getOperand1() {
         return operand1;
     }
 
-    public void setOperand1(Integer operand1) {
+    public void setOperand1(Double operand1) {
         this.operand1 = operand1;
     }
 
-    public Integer getOperand2() {
+    public Double getOperand2() {
         return operand2;
     }
 
-    public void setOperand2(Integer operand2) {
+    public void setOperand2(Double operand2) {
         this.operand2 = operand2;
     }
 
-    public Integer getResult() {
+    public Double getResult() {
         return result;
     }
 
 
     @Override
-    public Integer execute() {
+    public Double execute() {
         if(operand1 == null || operand2 == null){
             throw new IllegalArgumentException("One operand is missed");
         }
@@ -40,7 +40,7 @@ public class DivisionCommand implements OperationCommand {
         return result;
     }
     @Override
-    public Integer unexecute() {
+    public Double unexecute() {
         return operand1;
     }
 }
